Fix wxButtonToolBar compilation after adding new wxFont ctor overload.
The new wxFont ctor added in r70445 resulted in ambiguity when using
wxFont(int, wxFontFamily, wxFontStyle, int) ctor mixing the type-safe values
for the 2nd and 3rd arguments with "untyped" int for the 4th one. Fix this by
avoiding the use of this ctor as it seems impossible to do it in any other
way.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@70450 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775