Move checks for m_art != NULL before code using m_art in wxAuiToolBar.
It doesn't make sense to check whether a pointer is NULL after using it, so do
the checks first.

In fact the m_art pointer probably can never be NULL at all as other code uses
it without checking and so these checks could probably be removed completely,
but keep them for now as I'm not totally sure if having NULL art provider is
really impossible.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@70748 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775