Fix wxRadioButtons without explicit wxRB_GROUP styles in wxGTK.
If two radio buttons are created consecutively, they should be part of the
same radio group, even if wxRB_GROUP hadn't been explicitly used for the first
of them.

Do this is for consistency with wxMSW and also because the alternative
behaviour doesn't make much sense.

Also stop iterating too far backwards when searching for the group to be used
for a new radio button and add comments explaining what's going on here.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@71091 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775