Fix uninitialized variable warning in wxMSW wxFileSystemWatcher code.
This warning was harmless as the variable was in fact always initialized if
the code where it was used was reached but g++ 4.6.1 is not smart enough to
see it, just as MSVC for which we already had a workaround. So initialize it
explicitly just to suppress the warning.

See #14459.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@71976 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775