Fix light magenta definition in the colour database.
Make it different from magenta itself, otherwise a colour constructed from
"magenta" string was returning "light magenta" as its name.

The exact value of this colour is not well-defined, another commonly used one
seems to be 0xff80ff, but in a lot of places it is used as a synonym for
fuchsia pink which is pretty unambiguously 0xff77ff.

Closes #15628.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_3_0_BRANCH@75112 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775