Implement wxGetOsVersion() and wxGetOsDescription() with NSProcessInfo.
NSProcessInfo is the recommended way to determine OS X OS version, but the operatingSystemVersion property is only available in OS X 10.10+. Because of that a fallback to Gestalt() is implemented.

The NSProcessInfo.operatingSystemVersionString in the form “Version 10.10.4 (Build 14E46)"	now used by wxGetOsDescription() should be more useful to the user than the carbon implementations darwin version.